Moved vocab entity framework code into an appropriate subdirectory.
