+++ GermanVocabApp.Api/DependencyInjection/IServiceCollectionInjectionExtensions.cs
@@ -7,7 +7,7 @@ using GermanVocabApp.Api.VocabLists.Models;
 using GermanVocabApp.Api.VocabLists.Validation;
 using GermanVocabApp.Core.Contracts;
 using GermanVocabApp.DataAccess.EntityFramework;
-using GermanVocabApp.DataAccess.EntityFramework.Repositories;
+using GermanVocabApp.DataAccess.EntityFramework.Vocab;
 using GermanVocabApp.DataAccess.Shared;
 using GermanVocabApp.DataAccess.Shared.DataTransfer;
 using Microsoft.EntityFrameworkCore;
