+++ GermanVocabApp.DataAccess.EntityFramework.Tests.Unit/InMemoryVocabDatabaseSeeder.cs
@@ -1,4 +1,4 @@
-﻿using GermanVocabApp.DataAccess.EntityFramework.Models;
+﻿using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
 using GermanVocabApp.DataAccess.Models.Builders;
 using Microsoft.EntityFrameworkCore;
