+++ GermanVocabApp.DataAccess.EntityFramework/Configuration/Entities/VocabListItemEntityConfiguration.cs
@@ -1,4 +1,4 @@
-﻿using GermanVocabApp.DataAccess.EntityFramework.Models;
+﻿using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
