+++ GermanVocabApp.DataAccess.EntityFramework/Vocab/Models/Builders/Boilerplate/VocabListBuilder.boilerplate.cs
@@ -1,4 +1,5 @@
-﻿using GermanVocabApp.DataAccess.EntityFramework.Models;
+﻿using GermanVocabApp.DataAccess.EntityFramework.Core;
+using GermanVocabApp.DataAccess.EntityFramework.Vocab.Models;
 
 namespace GermanVocabApp.DataAccess.Models.Builders;
